+++ src/components/Form/Picklist.js
@@ -196,7 +196,7 @@ PicklistRaw.defaultProps = {
   items: [],
   labelInput: '',
   labelMultiselect: '',
-  size: 'small',
+  size: null,
 };
 
 export default enhanceWithClickOutside(PicklistRaw);