+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
-## Unreleased
+## 5.1.0
 
-* `Picklist` default size is now full-width, was `small`
+* New: `ScoreIcon` component
+* New: `Picklist` can now auto-close when selecting an item
+* Changed: `Picklist` default size is now full-width, was `small`
+* Fixed: `ObjectHome` doesn't show the menu style if no menu is provided
 
 ## 5.0.0